ui: shutdown screen and split out dashboard components
