+++ src/Views/Dashboard/PrintingDashboardView.swift
@@ -1,0 +1,21 @@
+//
+//  PrintingDashboardView.swift
+//  Plastic
+//
+//  Created by Charles Pickering on 9/19/22.
+//
+
+import SwiftUI
+
+struct PrintingDashboardView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct PrintingDashboardView_Previews: PreviewProvider {
+    static var previews: some View {
+        PrintingDashboardView()
+            .environmentObject(AppEnv())
+    }
+}
